manager log in good but dashboard not

+++ database/migrations/2025_05_02_134416_modify_sessions_table_for_string_user_id.php
@@ -6,25 +6,17 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up()
     {
         Schema::table('sessions', function (Blueprint $table) {
             $table->string('user_id')->nullable()->change();
         });
     }
-    
+
     public function down()
     {
         Schema::table('sessions', function (Blueprint $table) {
-            $table->integer('user_id')->nullable()->change();
+            $table->unsignedBigInteger('user_id')->nullable()->change();
         });
     }
-
-    /**
-     * Reverse the migrations.
-     */
-   
 };
